# deployment-strategies/foundry_agentic_app/typescript-foundry-agent-service-aca/contracts/backend-api.openapi.yaml
...
openapi: 3.1.0
info:
  title: CAIRA Business API
  version: 2.0.0
  description: |
    REST contract for the CAIRA business API container. This is the public-facing
    API that the frontend calls. It wraps the agent container's generic conversation
    API with a business-oriented business layer.

    The API container is agent-framework-agnostic — it only knows the agent API
    contract (`agent-api.openapi.yaml`) and proxies requests to whichever agent
    implementation is deployed.

    ## Business Operations

    Every conversation starts from a **business operation** — there is no generic
    "create a conversation" endpoint. The three operations are:

    - `POST /api/activities/discovery` — Start opportunity discovery
    - `POST /api/activities/planning` — Start account planning
    - `POST /api/activities/staffing` — Start team staffing

    Each operation creates a conversation on the agent container, sends a synthetic
    first message aligned to the selected activity, and returns the conversation ID
    and the agent's opening response.

    ## Activity Resolution

    Each activity has a **definitive conclusion** captured via agent-driven resolution
    tools (`resolve_discovery`, `resolve_planning`, `resolve_staffing`). When an activity
    resolves, the SSE stream includes an `activity.resolved` event with structured
    outcome data. The API captures this outcome and stores it on the conversation.

...
    the outcome is preserved.

    ## Streaming

...

...

...

    ```
...
servers:
...
paths:
  /api/activities/discovery:
    post:
      operationId: startDiscovery
      summary: Start opportunity discovery
      tags: [Activities]
      description: |
        Starts an opportunity discovery flow. Creates a conversation on the
        agent container, sends a synthetic first message for discovery routing,
        and returns the conversation ID with the agent's opening response.
      responses:
        '201':
          description: Discovery flow started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityConversationStarted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '502':
          $ref: '#/components/responses/AgentUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/activities/planning:
    post:
      operationId: startPlanning
      summary: Start account planning
      tags: [Activities]
      description: |
        Starts an account planning flow. The agent presents planning options,
        evaluates the selected path, and returns the conversation ID with the
        opening response.
      responses:
        '201':
          description: Planning flow started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityConversationStarted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '502':
          $ref: '#/components/responses/AgentUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/activities/staffing:
    post:
      operationId: startStaffing
      summary: Start team staffing
      tags: [Activities]
      description: |
        Starts a team staffing interview. The agent evaluates the requested
        engagement needs, recommends staffing coverage, and returns the
        conversation ID with the opening response.
      responses:
        '201':
          description: Staffing interview started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityConversationStarted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '502':
          $ref: '#/components/responses/AgentUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/activities/conversations:
    get:
      operationId: listActivityConversations
      summary: List all conversations
      tags: [ActivityConversations]
      description: |
        Returns a paginated list of conversations (conversations), ordered by most
        recent activity. Each conversation includes its mode, status, and outcome
        (if resolved). Maps to `GET /conversations` on the agent container with
        business-layer enrichment.
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated conversation list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityConversationList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '502':
          $ref: '#/components/responses/AgentUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/activities/conversations/{conversationId}:
    get:
      operationId: getActivityConversation
      summary: Get conversation details
      tags: [ActivityConversations]
      description: |
        Returns an activity conversation's details including mode, status, outcome (if resolved),
        and full message (message) history. Maps to `GET /conversations/{conversationId}`
        on the agent container with business-layer enrichment.
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          description: ActivityConversation with message history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityConversationDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/AgentUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/activities/conversations/{conversationId}/messages:
    post:
      operationId: message
      summary: Send a message in an activity conversation
      tags: [ActivityConversations]
      description: |
        Sends a message to the agent within an existing conversation and receives
        a response. The response format depends on the `Accept` header:

        - `Accept: text/event-stream` (default) — streams the assistant response
          as SSE events. The API parses the agent stream to detect `activity.resolved`
          events, captures the outcome, and passes all events through to the frontend.
        - `Accept: application/json` — waits for the full response

        Maps to `POST /conversations/{conversationId}/messages` on the agent container.

        ### SSE Event Schema

        **`message.delta`**
        ```json
        {"content": "partial assistant text"}
        ```

        **`activity.resolved`**
        ```json
        {"tool": "resolve_discovery", "result": {"fit": "qualified", "signals_reviewed": 4, "primary_need": "..."}}
        ```

        **`message.complete`**
        ```json
        {
          "messageId": "msg_abc123",
          "content": "That is a strong next question.",
          "usage": {
            "promptTokens": 10,
            "completionTokens": 12
          }
        }
        ```

        **`error`**
        ```json
        {"code": "agent_error", "message": "The service could not complete the request"}
        ```
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
      responses:
        '200':
          description: |
            Activity response. Content type depends on the `Accept` header:
            - `text/event-stream` for SSE streaming
            - `application/json` for complete response
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE event stream. See operation description for event format.
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/AgentUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/activities/stats:
    get:
      operationId: getStats
      summary: Get activity statistics
      tags: [Observability]
      description: |
        Returns aggregate statistics about activity — total conversations
        per mode, resolution counts, and active conversations.
      responses:
        '200':
          description: Activity statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityStats'
        '400':
          $ref: '#/components/responses/BadRequest'
        '502':
          $ref: '#/components/responses/AgentUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bffBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token injected by the frontend BFF for internal service-to-service authentication.
  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      description: Unique identifier of the activity conversation
      schema:
        type: string
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    ActivityMode:
      type: string
      description: The type of activity
      enum:
        - discovery
        - planning
        - staffing
    ActivityStatus:
      type: string
      description: Whether the activity has reached its conclusion
      enum:
        - active
        - resolved
    ActivityConversationStarted:
      type: object
      description: |
        Response from a business operation that starts an conversation. Contains
        the conversation ID, mode, initial status, and the synthetic user message
        that the frontend should display and send via the streaming message
        endpoint to get the agent's first response.
      required:
        - id
        - mode
        - status
        - syntheticMessage
        - createdAt
      properties:
        id:
          type: string
          description: Unique conversation identifier (same as conversation ID)
        mode:
          $ref: '#/components/schemas/ActivityMode'
        status:
          $ref: '#/components/schemas/ActivityStatus'
        syntheticMessage:
          type: string
          description: >
            The synthetic user message for this conversation mode. The frontend
            should display this as the first user bubble and send it via
            the message streaming endpoint to obtain the agent's opening response.
        createdAt:
          type: string
          format: date-time
          description: When the conversation was created
...
    ActivityConversationDetail:
      type: object
      description: ActivityConversation with full message history
      required:
        - id
        - mode
        - status
        - createdAt
        - lastMessageAt
        - messageCount
        - messages
      properties:
        id:
          type: string
          description: Unique conversation identifier
        mode:
          $ref: '#/components/schemas/ActivityMode'
        status:
          $ref: '#/components/schemas/ActivityStatus'
        outcome:
          $ref: '#/components/schemas/ActivityOutcome'
        createdAt:
          type: string
          format: date-time
          description: When the conversation was created
        lastMessageAt:
          type: string
          format: date-time
          description: When the last message occurred
        messageCount:
          type: integer
          description: Total number of messages
        messages:
          type: array
          maxItems: 1000
          items:
            $ref: '#/components/schemas/ActivityMessage'
          description: Messages in chronological order
...
    ActivityConversationList:
      type: object
      description: Paginated list of conversations
      required:
        - conversations
        - offset
        - limit
        - total
      properties:
        conversations:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/ActivityConversation'
          description: ActivityConversations on this page
        offset:
          type: integer
          description: Number of items skipped
        limit:
          type: integer
          description: Maximum items per page
        total:
          type: integer
          description: Total number of conversations
    MessageRequest:
      type: object
      description: A message to send to the assistant
      required:
        - message
      properties:
        message:
          type: string
          minLength: 1
          description: The user's message text
...
  responses:
    BadRequest:
      description: Invalid request (malformed body, missing required fields)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: ActivityConversation not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Too many requests — upstream LLM rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    AgentUnavailable:
      description: Agent container is unavailable or returned an error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - bffBearerAuth: []